requirements.txt
category_encoders==2.3.0
datasets==1.16.1
matplotlib==3.3.2
nlpaug==1.1.9
nltk==3.5
numpy==1.20.2
pandas==1.3.5
Pillow==9.1.0
PyYAML==6.0
scikit_learn==1.0.2
scipy==1.5.2
seaborn==0.11.0
torch==1.7.1
torchvision==0.8.2
tqdm==4.62.3
transformers==4.14.1
xgboost==1.4.2
